Get rid of code we aren't using (and are breaking).
